Refactored parameter file selection
