+++ src/vscode-bicep/src/deploy/selectParameterFile.ts
@@ -1,0 +1,76 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+import * as path from "path";
+import vscode, { Uri } from "vscode";
+import {
+  IActionContext,
+  IAzureQuickPickItem
+} from "@microsoft/vscode-azext-utils";
+import { appendToOutputChannel } from "../utils/logger";
+
+export async function selectParameterFile(
+  _context: IActionContext,
+  sourceUri: Uri | undefined
+): Promise<string> {
+  const quickPickList: IQuickPickList =
+    await createParameterFileQuickPickList();
+  const result: IAzureQuickPickItem<IPossibleParameterFile | undefined> =
+    await _context.ui.showQuickPick(quickPickList.items, {
+      canPickMany: false,
+      placeHolder: `Select a parameter file`,
+      suppressPersistence: true,
+    });
+
+  if (result === quickPickList.browse) {
+    const paramsPaths: Uri[] | undefined = await vscode.window.showOpenDialog({
+      canSelectMany: false,
+      defaultUri: sourceUri,
+      openLabel: "Select Parameter File",
+    });
+    if (paramsPaths && paramsPaths.length == 1) {
+      const parameterFilePath = paramsPaths[0].fsPath;
+      appendToOutputChannel(
+        `Parameter file used in deployment -> ${path.basename(parameterFilePath)}`
+      );
+      return parameterFilePath;
+    }
+  }
+
+  appendToOutputChannel(`Parameter file was not provided`);
+
+  return "";
+}
+
+async function createParameterFileQuickPickList(): Promise<IQuickPickList> {
+  const none: IAzureQuickPickItem<IPossibleParameterFile | undefined> = {
+    label: "$(circle-slash) None",
+    data: undefined,
+  };
+  const browse: IAzureQuickPickItem<IPossibleParameterFile | undefined> = {
+    label: "$(file-directory) Browse...",
+    data: undefined,
+  };
+
+  const pickItems: IAzureQuickPickItem<IPossibleParameterFile | undefined>[] = [
+    none,
+  ].concat([browse]);
+
+  return {
+    items: pickItems,
+    none,
+    browse,
+  };
+}
+
+interface IQuickPickList {
+  items: IAzureQuickPickItem<IPossibleParameterFile | undefined>[];
+  none: IAzureQuickPickItem<IPossibleParameterFile | undefined>;
+  browse: IAzureQuickPickItem<IPossibleParameterFile | undefined>;
+}
+
+interface IPossibleParameterFile {
+  uri: Uri;
+  friendlyPath: string;
+  isCloseNameMatch: boolean;
+  fileNotFound?: boolean;
+}
